psychmeta version 0.2.4
o A new method to account for sampling error in interactive artifact
distributions has been added (NOTE: this is now the default methods
used in interactive AD meta-analyses).
o Stability improvements for dMod functions.
o Artifacts harvested from rows of a database that do not contain
effects sizes in the ma_r and ma_d functions are now organized to avoid
redundancy with observations from the same sample that are included in
the meta-analytic computations.
o Bug fixes to metabulate().