Updated handling of artifact distributions in ma_r
The artifact information from pre-specified distributions and from observations not included in a meta-analysis are now fully supported for pairwise AD and moderated AD approaches. This means that the supplemental_ads argument is no longer restricted to a list format - it can now have a tibble format in which construct-pair and moderator breakdowns are specified and named accordingly to the meta-analysis information.