Changed "treat_as_d" argument to "treat_as_r"
Changed "treat_as_d" argument to ma_d and ma_d_ic to "treat_as_r", as "treat_as_r" is a more general name that could be used in future meta-analysis functions of effect sizes in non-correlation metrics. This will not disrupt analyses using function calls that include treat_as_d, as that is still recognized by the meta-analysis functions as an ellipsis argument.