Bug fixes for bb d analyses from ic analyses
Updated initial conversion of bare-bones meta-analytic results to correlations within ma_r_ic. The conversion now to uses the meta-analytic weights (rather than n_adj) to determine the mean proportion value to use in the conversion - this makes the initial conversion a mirror to the back-conversion that takes place within convert_ma.